+++ CHANGELOG.md
@@ -38,3 +38,5 @@ New function `derivSpray`, to differentiate a spray.
 * New function `sprayDivision`, to perform the division of a spray by a list of sprays.
 
 * New function `groebner`, to compute a Groebner basis of a list of sprays.
+
+* New function `isSymmetricSpray`, to check whether a spray is a symmetric polynomial.